Update Sound-Clock.ino
The clock was not staying in WAITING state, it was just going directly to what sounded like SPEAK_MINUTE.  In fact, it wasn't in SPEAK_MINUTE, it was going through SPEAK_HOUR so quickly that we weren't hearing it.  The bug was on line 120 where the default return value from DoWaiting() should have been to stay in WAITING, but we were returning SPEAK_HOUR in every case, so it never actually waited.