Sound playback MUCH IMPROVED!
This is kind of dumb, because I copied all of the samples over into the test directory, but I just want to check this in right now... it was hard to find the one thing that had to change.  Biggest factor was NOT resetting the sample rate when loading a new sound sample.